@@ -27,7 +27,7 @@
     "preferred-install": "dist"
   },
   "require": {
-    "php": ">=7.1.0",
+    "php": "^8.0",
     "ext-json": "*"
   },
   "autoload": {
